move planting site stats to discrete file
